+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'ddangddangMarket'
+rootProject.name = 'bidderbidder'
