+++ src/main/java/com/fakedevelopers/bidderbidder/domain/Constants.java
@@ -1,0 +1,5 @@
+package com.fakedevelopers.bidderbidder.domain;
+
+public class Constants {
+    public static final int ITEMMAXCOUNT = 10000;
+}
