+++ src/main/java/com/fakedevelopers/bidderbidder/dto/PageListResponseDto.java
@@ -1,0 +1,15 @@
+package com.fakedevelopers.bidderbidder.dto;
+
+import lombok.Getter;
+import java.util.List;
+
+@Getter
+public class PageListResponseDto {
+    private final int itemCount;
+    private final List<ProductListDto> items;
+
+    public PageListResponseDto(int itemCount, List<ProductListDto> items) {
+        this.itemCount = itemCount;
+        this.items = items;
+    }
+}
