+++ src/main/java/com/fakedevelopers/bidderbidder/dto/ProductListDto.java
@@ -1,0 +1,19 @@
+package com.fakedevelopers.bidderbidder.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ProductListDto {
+
+    private final long boardId;
+    private final String thumbnail;
+    private final String boardTitle;
+    private final Long hopePrice;
+    private final long openingBid;
+    private final long tick;
+    private final String expirationDate;
+    private final int bidderCount;
+
+}
