+++ src/main/java/com/fakedevelopers/bidderbidder/exception/InvalidCategoryException.java
@@ -1,0 +1,7 @@
+package com.fakedevelopers.bidderbidder.exception;
+
+public class InvalidCategoryException extends RuntimeException{
+    public InvalidCategoryException(String message){
+        super(message);
+    }
+}
