+++ src/main/java/com/fakedevelopers/bidderbidder/exception/InvalidExpirationDateException.java
@@ -1,0 +1,7 @@
+package com.fakedevelopers.bidderbidder.exception;
+
+public class InvalidExpirationDateException extends RuntimeException{
+    public InvalidExpirationDateException(String message){
+        super(message);
+    }
+}
