+++ src/main/java/com/fakedevelopers/bidderbidder/exception/InvalidExtensionException.java
@@ -1,0 +1,7 @@
+package com.fakedevelopers.bidderbidder.exception;
+
+public class InvalidExtensionException extends RuntimeException{
+    public InvalidExtensionException(String message){
+        super(message);
+    }
+}
