+++ src/main/java/com/fakedevelopers/bidderbidder/exception/InvalidHopePriceException.java
@@ -1,0 +1,7 @@
+package com.fakedevelopers.bidderbidder.exception;
+
+public class InvalidHopePriceException extends RuntimeException {
+    public InvalidHopePriceException(String message) {
+        super(message);
+    }
+}
