+++ src/main/java/com/fakedevelopers/bidderbidder/exception/InvalidRepresentPictureIndexException.java
@@ -1,0 +1,7 @@
+package com.fakedevelopers.bidderbidder.exception;
+
+public class InvalidRepresentPictureIndexException extends RuntimeException {
+    public InvalidRepresentPictureIndexException(String message) {
+        super(message);
+    }
+}
