+++ src/main/java/com/fakedevelopers/bidderbidder/model/Category.java
@@ -1,0 +1,10 @@
+package com.fakedevelopers.bidderbidder.model;
+
+// 상의 후 제대로 정하자
+public enum Category {
+    FOOD,
+    COOK,
+    TOY;
+
+    public static final Category[] VALUES = Category.values();
+}
