+++ src/main/java/com/fakedevelopers/bidderbidder/service/UserService.java
@@ -1,0 +1,7 @@
+package com.fakedevelopers.bidderbidder.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+}
