--- src/main/java/com/fakedevelopers/ddangddangmarket/AuctionProjectApplication.java
@@ -1,15 +1,0 @@
-package com.fakedevelopers.ddangddangmarket;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@SpringBootApplication
-public class AuctionProjectApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AuctionProjectApplication.class, args);
-    }
-
-}
