--- src/main/java/com/fakedevelopers/ddangddangmarket/domain/Constants.java
@@ -1,5 +1,0 @@
-package com.fakedevelopers.ddangddangmarket.domain;
-
-public class Constants {
-    public static final int ITEMMAXCOUNT = 10000;
-}
