--- src/main/java/com/fakedevelopers/ddangddangmarket/exception/InvalidCategoryException.java
@@ -1,7 +1,0 @@
-package com.fakedevelopers.ddangddangmarket.exception;
-
-public class InvalidCategoryException extends RuntimeException{
-    public InvalidCategoryException(String message){
-        super(message);
-    }
-}
