--- src/main/java/com/fakedevelopers/ddangddangmarket/exception/InvalidExpirationDateException.java
@@ -1,7 +1,0 @@
-package com.fakedevelopers.ddangddangmarket.exception;
-
-public class InvalidExpirationDateException extends RuntimeException{
-    public InvalidExpirationDateException(String message){
-        super(message);
-    }
-}
