--- src/main/java/com/fakedevelopers/ddangddangmarket/exception/InvalidExtensionException.java
@@ -1,7 +1,0 @@
-package com.fakedevelopers.ddangddangmarket.exception;
-
-public class InvalidExtensionException extends RuntimeException{
-    public InvalidExtensionException(String message){
-        super(message);
-    }
-}
