--- src/main/java/com/fakedevelopers/ddangddangmarket/exception/InvalidHopePriceException.java
@@ -1,7 +1,0 @@
-package com.fakedevelopers.ddangddangmarket.exception;
-
-public class InvalidHopePriceException extends RuntimeException {
-    public InvalidHopePriceException(String message) {
-        super(message);
-    }
-}
