--- src/main/java/com/fakedevelopers/ddangddangmarket/model/Category.java
@@ -1,10 +1,0 @@
-package com.fakedevelopers.ddangddangmarket.model;
-
-// 상의 후 제대로 정하자
-public enum Category {
-    FOOD,
-    COOK,
-    TOY;
-
-    public static final Category[] VALUES = Category.values();
-}
