--- src/main/java/com/fakedevelopers/ddangddangmarket/service/UserService.java
@@ -1,7 +1,0 @@
-package com.fakedevelopers.ddangddangmarket.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-}
