+++ src/test/java/com/fakedevelopers/bidderbidder/AuctionProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.fakedevelopers.bidderbidder;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AuctionProjectApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
