--- src/test/java/com/fakedevelopers/ddangddangmarket/AuctionProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.fakedevelopers.ddangddangmarket;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuctionProjectApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
